+++ poll-taquito/src/App.tsx
@@ -1,17 +1,11 @@
 import * as React from "react";
-import AppBar from "@material-ui/core/AppBar";
-import Box from "@material-ui/core/Box";
-import Toolbar from "@material-ui/core/Toolbar";
-import Container from "@material-ui/core/Container";
-import Typography from "@material-ui/core/Typography";
-import Button from "@material-ui/core/Button";
 import { useBeaconWallet, useWallet } from "@tz-contrib/react-wallet-provider";
 import { initPollContract, initTezos, setWalletProvider } from "./contract";
-import CreatePollCard from "./CreatePollCard";
-import { Grid } from "@material-ui/core";
-import AddVoterCard from "./AddVoterCard";
-import RemoveVoterCard from "./RemoveVoterCard";
-import VoteCard from "./VoteCard";
+import CreatePollCard from "./components/CreatePollCard";
+import AddVoterCard from "./components/AddVoterCard";
+import RemoveVoterCard from "./components/RemoveVoterCard";
+import VoteCard from "./components/VoteCard";
+import Nav from "./components/Nav";
 import {
   BrowserRouter as Router,
   Switch,
@@ -35,45 +29,13 @@ function App() {
   }, [beaconWallet]);
   return (
     <Router>
-      <Box sx={{ flexGrow: 1 }}>
-        <AppBar position="static">
-          <Toolbar>
-            <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-              Hicvote
-            </Typography>
-            {activeAccount ? (
-              <>
-                {activeAccount.address}
-                <Button color="inherit" onClick={disconnect}>
-                  Logout
-                </Button>
-              </>
-            ) : (
-              <Button color="inherit" onClick={connect}>
-                Login
-              </Button>
-            )}
-          </Toolbar>
-        </AppBar>
+        <Nav/>
         <Switch>
           <Route exact path="/">
             <main>
-              <Container sx={{ marginTop: "2rem" }}>
-                <Grid container direction="row" spacing={3}>
-                  <Grid item xs={6}>
-                    <CreatePollCard />
-                  </Grid>
-                  <Grid item xs={6}>
-                    <VoteCard />
-                  </Grid>
-                  <Grid item xs={6}>
-                    <AddVoterCard />
-                  </Grid>
-                  <Grid item xs={6}>
-                    <RemoveVoterCard />
-                  </Grid>
-                </Grid>
-              </Container>
+              <Link to="/vote/1">Vote on Poll 1</Link><br></br>
+              <Link to="/vote/2">Vote on Poll 2</Link><br></br>
+              <Link to="/vote/3">Vote on Poll 3</Link>
             </main>
           </Route>
           <Route path="/about">
@@ -85,8 +47,12 @@ function App() {
           <Route path="/vote/:poll">
             <VoteCard/>
           </Route>
+          <Route path="/admin">
+            <CreatePollCard />
+            <AddVoterCard />
+            <RemoveVoterCard />
+          </Route>
         </Switch>
-      </Box>
     </Router>
   );
 }
