+++ poll-taquito/src/components/CreatePollCard.tsx
@@ -3,27 +3,32 @@ import * as Yup from "yup";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 import CardContent from "@material-ui/core/CardContent";
+import DatePicker from "@material-ui/lab/DatePicker";
 import { Formik, Form, Field } from "formik";
-import { Button, Grid } from "@material-ui/core";
+import { Button, Grid, TextField } from "@material-ui/core";
 import { useWallet } from "@tz-contrib/react-wallet-provider";
-import { vote } from "./contract";
+import { createPoll } from "../contract";
 import { useToasts } from "react-toast-notifications";
 import { FormikTextField } from "./FormikTextField";
-import { useParams } from "react-router-dom";
 
-export default function VoteCard() {
-  const poll = useParams();
-  console.log(poll);
+export default function CreatePollCard() {
   const { connected } = useWallet();
   const { addToast } = useToasts();
   const validationSchema = Yup.object().shape({
     pollId: Yup.string().required("Required"),
-    pollOption: Yup.number().min(1, "Min value is 1").required("Required"),
+    endDate: Yup.date().required("Required"),
+    noOfOptions: Yup.number()
+      .min(2, "Min 2 options required")
+      .required("Required"),
   });
   const handleSubmit = async (values: any, helper: any) => {
     if (connected) {
       try {
-        const hash = await vote(values.pollId, values.pollOption);
+        const hash = await createPoll(
+          values.pollId,
+          values.endDate,
+          values.noOfOptions
+        );
         if (hash) {
           addToast("Tx Submitted", {
             appearance: "success",
@@ -43,40 +48,47 @@ export default function VoteCard() {
   };
   return (
     <Card sx={{ maxWidth: "21.5rem" }}>
-      <CardHeader title="Cast your Vote" subheader="Cast your vote here" />
+      <CardHeader title="Create A Poll" subheader="Start a new poll" />
       <CardContent>
         <Formik
-          initialValues={{ pollId: "", pollOption: 1 }}
+          initialValues={{ pollId: "", endDate: new Date(), noOfOptions: 2 }}
           onSubmit={handleSubmit}
           validationSchema={validationSchema}
-          validateOnChange
           validateOnBlur
+          validateOnChange
         >
-          {({ setFieldValue, errors, touched, isValid, dirty }) => (
+          {({ setFieldValue, errors, values, touched, isValid, dirty }) => (
             <Form>
               <Grid direction="column" container spacing={3}>
-                { Object.keys(poll).length !== 0 ? (
-                  <Field
-                    name="pollId"
-                    type="hidden"
-                    label="Poll ID"
-                    value={ poll }
-                  />
-                ) : (
-                  <Grid item>
-                    <Field
-                      component={FormikTextField}
-                      name="pollId"
-                      type="text"
-                      label="Poll ID"
-                      fullWidth
-                    />
-                  </Grid>
-                )}
                 <Grid item>
                   <Field
                     component={FormikTextField}
-                    name="pollOption"
+                    name="pollId"
+                    type="text"
+                    label="Poll ID"
+                    fullWidth
+                  />
+                </Grid>
+                <Grid item>
+                  <Field
+                    label="End Date"
+                    name="endDate"
+                    component={DatePicker}
+                    onChange={(value: any) => {
+                      setFieldValue("endDate", value);
+                    }}
+                    renderInput={(params: any) => (
+                      <TextField {...params} fullWidth />
+                    )}
+                    error={touched.endDate && Boolean(errors.endDate)}
+                    helperText={touched.endDate ? errors.endDate : ""}
+                    disablePast
+                  />
+                </Grid>
+                <Grid item>
+                  <Field
+                    component={FormikTextField}
+                    name="noOfOptions"
                     type="number"
                     label="Number of options"
                     fullWidth
