+++ poll-taquito/src/components/VoteCard.tsx
@@ -3,32 +3,27 @@ import * as Yup from "yup";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 import CardContent from "@material-ui/core/CardContent";
-import DatePicker from "@material-ui/lab/DatePicker";
 import { Formik, Form, Field } from "formik";
-import { Button, Grid, TextField } from "@material-ui/core";
+import { Button, Grid } from "@material-ui/core";
 import { useWallet } from "@tz-contrib/react-wallet-provider";
-import { createPoll } from "./contract";
+import { vote } from "../contract";
 import { useToasts } from "react-toast-notifications";
 import { FormikTextField } from "./FormikTextField";
+import { useParams } from "react-router-dom";
 
-export default function CreatePollCard() {
+export default function VoteCard() {
+  const poll = useParams();
+  console.log(poll);
   const { connected } = useWallet();
   const { addToast } = useToasts();
   const validationSchema = Yup.object().shape({
     pollId: Yup.string().required("Required"),
-    endDate: Yup.date().required("Required"),
-    noOfOptions: Yup.number()
-      .min(2, "Min 2 options required")
-      .required("Required"),
+    pollOption: Yup.number().min(1, "Min value is 1").required("Required"),
   });
   const handleSubmit = async (values: any, helper: any) => {
     if (connected) {
       try {
-        const hash = await createPoll(
-          values.pollId,
-          values.endDate,
-          values.noOfOptions
-        );
+        const hash = await vote(values.pollId, values.pollOption);
         if (hash) {
           addToast("Tx Submitted", {
             appearance: "success",
@@ -48,47 +43,40 @@ export default function CreatePollCard() {
   };
   return (
     <Card sx={{ maxWidth: "21.5rem" }}>
-      <CardHeader title="Create A Poll" subheader="Start a new poll" />
+      <CardHeader title="Cast your Vote" subheader="Cast your vote here" />
       <CardContent>
         <Formik
-          initialValues={{ pollId: "", endDate: new Date(), noOfOptions: 2 }}
+          initialValues={{ pollId: "", pollOption: 1 }}
           onSubmit={handleSubmit}
           validationSchema={validationSchema}
-          validateOnBlur
           validateOnChange
+          validateOnBlur
         >
-          {({ setFieldValue, errors, values, touched, isValid, dirty }) => (
+          {({ setFieldValue, errors, touched, isValid, dirty }) => (
             <Form>
               <Grid direction="column" container spacing={3}>
-                <Grid item>
+                { Object.keys(poll).length !== 0 ? (
                   <Field
-                    component={FormikTextField}
                     name="pollId"
-                    type="text"
+                    type="hidden"
                     label="Poll ID"
-                    fullWidth
+                    value={ poll }
                   />
-                </Grid>
-                <Grid item>
-                  <Field
-                    label="End Date"
-                    name="endDate"
-                    component={DatePicker}
-                    onChange={(value: any) => {
-                      setFieldValue("endDate", value);
-                    }}
-                    renderInput={(params: any) => (
-                      <TextField {...params} fullWidth />
-                    )}
-                    error={touched.endDate && Boolean(errors.endDate)}
-                    helperText={touched.endDate ? errors.endDate : ""}
-                    disablePast
-                  />
-                </Grid>
+                ) : (
+                  <Grid item>
+                    <Field
+                      component={FormikTextField}
+                      name="pollId"
+                      type="text"
+                      label="Poll ID"
+                      fullWidth
+                    />
+                  </Grid>
+                )}
                 <Grid item>
                   <Field
                     component={FormikTextField}
-                    name="noOfOptions"
+                    name="pollOption"
                     type="number"
                     label="Number of options"
                     fullWidth
